refactor: godoc examples in separate test files
