+++ influxdb3/query_test.go
@@ -24,7 +24,6 @@ package influxdb3
 
 import (
 	"context"
-	"log"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -54,60 +53,4 @@ func TestQueryWithOptionsNotSet(t *testing.T) {
 	assert.Nil(t, iterator)
 	assert.Error(t, err)
 	assert.EqualError(t, err, "options not set")
-}
-
-func ExampleClient_Query() {
-	client, err := NewFromEnv()
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer client.Close()
-
-	// query
-	iterator, err := client.Query(context.Background(),
-		"SELECT count(*) FROM weather WHERE time >= now() - interval '5 minutes'")
-
-	for iterator.Next() {
-		// process the result
-	}
-
-	// query with custom header
-	iterator, err = client.Query(context.Background(),
-		"SELECT count(*) FROM stat WHERE time >= now() - interval '5 minutes'",
-		WithHeader("X-trace-ID", "#0122"))
-
-	for iterator.Next() {
-		// process the result
-	}
-}
-
-func ExampleClient_QueryWithParameters() {
-	client, err := NewFromEnv()
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer client.Close()
-
-	// query
-	iterator, err := client.QueryWithParameters(context.Background(),
-		"SELECT count(*) FROM weather WHERE location = $location AND time >= now() - interval '5 minutes'",
-		QueryParameters{
-			"location": "sun-valley-1",
-		})
-
-	for iterator.Next() {
-		// process the result
-	}
-
-	// query with custom header
-	iterator, err = client.QueryWithParameters(context.Background(),
-		"SELECT count(*) FROM weather WHERE location = $location AND time >= now() - interval '5 minutes'",
-		QueryParameters{
-			"location": "sun-valley-1",
-		},
-		WithHeader("X-trace-ID", "#0122"))
-
-	for iterator.Next() {
-		// process the result
-	}
 }
